+++ test/src/test_misc.cpp
@@ -461,6 +461,35 @@ void test_wait_until(UnitTest& ut)
     CUN_UNITTEST_RESET(ut);
 }
 
+void test_wait_while(UnitTest& ut)
+{
+    CUN_UNITTEST_TITLE(ut, "Test code: Miscellaneous - wait_while.");
+    CUN_UNITTEST_NL(ut);
+
+    CUN_UNITTEST_EXEC(ut, int count);
+    CUN_UNITTEST_NL(ut);
+
+    CUN_UNITTEST_NAME(ut, "try 1 times");
+    CUN_UNITTEST_EXEC(ut, count = 0);
+    CUN_UNITTEST_EXEC(ut, wait_while([&count]{ ++count; return false; }));
+    CUN_UNITTEST_EVAL(ut, count == 1);
+    CUN_UNITTEST_NL(ut);
+
+    CUN_UNITTEST_NAME(ut, "try 2 times");
+    CUN_UNITTEST_EXEC(ut, count = 0);
+    CUN_UNITTEST_EXEC(ut, wait_while([&count]{ ++count; return count < 2; }));
+    CUN_UNITTEST_EVAL(ut, count == 2);
+    CUN_UNITTEST_NL(ut);
+
+    CUN_UNITTEST_NAME(ut, "try 10 times");
+    CUN_UNITTEST_EXEC(ut, count = 0);
+    CUN_UNITTEST_EXEC(ut, wait_while([&count]{ ++count; return count < 10; }));
+    CUN_UNITTEST_EVAL(ut, count == 10);
+    CUN_UNITTEST_NL(ut);
+
+    CUN_UNITTEST_RESET(ut);
+}
+
 } // namespace
 
 int main()
@@ -479,6 +508,7 @@ int main()
     test_remap_range(ut);
     test_basename(ut);
     test_wait_until(ut);
+    test_wait_while(ut);
 
     return EXIT_SUCCESS;
 }