+++ ChangeLog.md
@@ -7,6 +7,7 @@ ChangeLog
 ### Added
 
 * Library: Miscellaneous: wait_until: The first implementation.
+* Library: Miscellaneous: wait_while: The first implementation.
 * Library: Software timer: The first implementation.
 
 ### Changed